requirements.txt
flask
flask-sqlalchemy
flask-admin
pyyaml
pyjwt